feat(scoring): Phase β2 ClusterScoringContext share-weighted fan-out (#11)
`_build_contexts` 를 winner-takes-all 에서 G×T×F cross-product
share-weighted fan-out 으로 재작성 (spec §2.4).

핵심 변경:
- N<3 (G/T/F 한 축이라도 비어있는) item 은 어떤 cluster 에도 기여 0
  (이전: winner cluster 에 0.x 기여). assign_shares 빈 dict 정책과 정합.
- N=3 item 의 G×T×F share 합 = 1.0 (mass preservation invariant).
- ClusterScoringContext schema 부분 완화: youtube_video_count int→float,
  post_count_today int→float. post_count_total 는 history schema 호환을
  위해 int 유지 (γ 에서 float 마이그 예정).
- accounts 는 winner-keyed 유지 (fan-out 으로 부풀려지면 안됨).
- post_count_total = history.get_total_post_count + round(share-sum) —
  γ 마이그 전 mixed precision 보존.
- KeyError 방지: grouped 의 모든 winner key 는 N<3 라도 zero-aggregate
  context 를 받음 (score_and_export 의 decisions[key] 룩업 안전).

Behavior change (PR body 에서 확장):
- cluster space 가 G×T×F cross-product 로 확장 — minmax_same_run 분모
  row 수가 증가. 같은 raw count 라도 normalized score 가 달라질 수 있음.
  γ 에서 numerator/denominator align 예정.
- avg_engagement_rate 분모: max(len(items), 1) → max(post_count_today, 1.0).
  fractional share 합이라 winner-takes-all 보다 inflated 될 수 있음.
- fan-out-only cluster (β2 acc 에 있지만 grouped 에 winner 가 없는 cluster)
  은 score 계산되지만 summary 는 β3 에서 wire 예정.

Scope (advisor 합의 locked):
- 건드림: pipelines/run_scoring_pipeline._build_contexts +
  scoring/cluster_context.ClusterScoringContext + 영향받은 unit tests.
- 미건드림: aggregation/build_cluster_summary.group_by_cluster (β3 영역),
  drilldown (β4 영역), score_history_weekly (γ 영역, history schema
  마이그 부담), representative_builder (이미 share-weighted).

Tests:
- tests/unit/test_run_scoring_pipeline_share_weighted.py (8 신규 pinning):
  N<3 zero contribution, N=3 mass preservation, share fan-out cross-product,
  youtube_video_count share-weighted, accounts winner-keyed, post_count_total
  history int + round(share-sum), empty grouped.
- 전체 721 passed (+ 8 신규 = 713 → 721).
- live smoke (2026-04-27 enriched 14 post): 14 cluster, top score=48.33,
  분포 정상.

+++ src/scoring/cluster_context.py
@@ -27,7 +27,7 @@ class ClusterScoringContext(BaseModel):
     social_weighted_engagement: float
 
     # --- YouTube (spec §9.2 YouTube) ---
-    youtube_video_count: int              # V, 최근 window_days 일
+    youtube_video_count: float            # V (Phase β2: share-weighted fan-out 으로 fractional)
     youtube_views_total: float            # 최근 window_days 일 조회수 합
     youtube_view_growth: float            # (this_week - last_week) / last_week. 0 안전처리.
 
@@ -41,7 +41,7 @@ class ClusterScoringContext(BaseModel):
     momentum_new_account_ratio: float     # 0~1.
 
     # --- 라이프사이클 / 방향성 / maturity 판정용 ---
-    post_count_total: int                 # 누적 (직전일까지 + 오늘)
-    post_count_today: int
+    post_count_total: int                 # 누적 (직전일까지 + 오늘). Phase γ TODO: history schema 마이그 후 float
+    post_count_today: float               # Phase β2: share-weighted fan-out 으로 fractional
     avg_engagement_rate: float
     # 전일 / 지난 주 같은 비교 기준은 orchestrator 가 별도 dict 로 넘긴다.
